// client/src/app/models/camera.ts
export interface Camera {
    id: number;
        name: string;
        brand: string;
        sensor: string;
        files: string;
        video: string[];
        lens: string;
        ios: string;
        shoot: string;
        dimensions: string;
        weight: number;
        features: string[];
        price: number;
        img: string;
}